% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.thrreg-package.R
\docType{package}
\name{fable.thrreg-package}
\alias{fable.thrreg}
\alias{fable.thrreg-package}
\title{fable.thrreg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
