% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabletools.R
\name{traverse}
\alias{traverse}
\title{Recursively traverse an object}
\usage{
traverse(
  x,
  .f = list,
  .g = identity,
  .h = identity,
  base = function(.x) rlang::is_syntactic_literal(.x) || rlang::is_symbol(.x)
)
}
\arguments{
\item{x}{The object to traverse}

\item{.f}{A function for combining the recursed components}

\item{.g}{A function applied to the object before recursion}

\item{.h}{A function applied to the base case}

\item{base}{The base case for the recursion}
}
\description{
Internal function in fabletools
}
\keyword{internal}
